Agents and Workflows/camel playground/requirements.txt
streamlit>=1.28.0
camel-ai>=0.2.57
plotly>=5.18.0
pandas>=2.1.0
python-dotenv>=1.0.0
openai>=1.3.0
anthropic>=0.3.11
matplotlib>=3.8.0
seaborn>=0.13.0
watchdog>=3.0.0
pillow>=10.1.0
rich>=13.7.0
google-generativeai>=0.3.0
requests>=2.31.0
beautifulsoup4>=4.12.0